% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_dew_point.R
\name{metar_dew_point}
\alias{metar_dew_point}
\title{Extract dew point temperature.}
\usage{
metar_dew_point(x)
}
\arguments{
\item{x}{Input character vector}
}
\value{
A numeric vector. A dew point temperature in degrees Celsius.
}
\description{
Function extract a dew point temperature value from METAR weather report.
}
\examples{
metar_dew_point("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_dew_point("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180")
metar_dew_point("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025 NOSIG= NOSIG=")

}
