% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_get.R
\name{metar_get}
\alias{metar_get}
\title{Get a current METAR report for an airport.}
\usage{
metar_get(airport = "EPWA")
}
\arguments{
\item{airport}{Input character vector with an ICAO airport code}
}
\value{
A character vector with a current METAR weather report.
}
\description{
Function gets metar report from a web page.\cr
\cr
If only airport is passed as an argument, a current METAR weather report is donwloaded from\cr
the web page https://aviationweather.gov/metar/\cr
\cr
}
\examples{
metar_get("EPWA")
metar_get("CYUL")
metar_get("LEMD")
metar_get("WAW")

}
