% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_hour.R
\name{metar_hour}
\alias{metar_hour}
\title{Extract hour and minutes.}
\usage{
metar_hour(x)
}
\arguments{
\item{x}{Input character vector}
}
\value{
A character vector HH:mm.
}
\description{
Function extract an hour and minutes from METAR weather report.
}
\examples{
metar_hour("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_hour("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180")
metar_hour("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025 NOSIG= NOSIG=")

}
