% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_iata_icao.R
\name{metar_iata_icao}
\alias{metar_iata_icao}
\title{Convert IATA airport code to ICAO airport code and vice versa.}
\usage{
metar_iata_icao(x)
}
\arguments{
\item{x}{Input character vector.}
}
\value{
A character vector.
}
\description{
Convert IATA airport code to ICAO airport code and vice versa.
}
\examples{
metar_iata_icao("EPWA")
metar_iata_icao("CYUL")
metar_iata_icao("LEMD")
metar_iata_icao("WAW")
metar_iata_icao("FRA")
metar_iata_icao("KRK")

}
