% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_rwy_visibility.R
\name{metar_rwy_visibility}
\alias{metar_rwy_visibility}
\title{Extract visibility.}
\usage{
metar_rwy_visibility(x, metric = TRUE)
}
\arguments{
\item{x}{Input character vector}

\item{metric}{logical; the default value is TRUE, a returned visibility will be in meters;
if FALSE then in feet.}
}
\value{
A numeric vector. A visibility in m/s or feet.
}
\description{
Function extracts runway(s) visibility value(s) from METAR weather report.
}
\examples{

metar_rwy_visibility("EBBR 040220Z VRB01KT 0150 R25L/1200N R02/P1500 07/06 Q1017")
metar_rwy_visibility("EBBR 040220Z VRB01KT 0150 R25R/0600FT R02/P1500 07/06 Q1017")
metar_rwy_visibility("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_rwy_visibility("CYWG 172000Z 30015G25KT 3/4SM R36/4000FT/D -SN M05/M08 A2992")
metar_rwy_visibility("EBBR 040220Z VRB01KT 0150 R25L/1200N R26R/1000 R36/4000FT/D -SN")

}
