% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_speed.R
\name{metar_speed}
\alias{metar_speed}
\title{Extract wind speed}
\usage{
metar_speed(x, metric = TRUE)
}
\arguments{
\item{x}{character; METAR weather report or reports.}

\item{metric}{For the default metric = TRUE a returned wind speed is in m/s. If it's FALSE, in knots.}
}
\value{
A numeric vector. A wind speed in m/s.
}
\description{
Function extracts a wind speed value from METAR weather report.
}
\examples{
metar_speed("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_speed("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180", metric = FALSE)
metar_speed("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025 NOSIG= NOSIG=")

}
