% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_time_zone.R
\name{metar_time_zone}
\alias{metar_time_zone}
\title{Extract time zone.}
\usage{
metar_time_zone(x)
}
\arguments{
\item{x}{Input character vector}
}
\value{
A character vector.
}
\description{
Function extracts a name of airport from METAR weather report.
}
\examples{
metar_time_zone("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_time_zone("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180")
metar_time_zone("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025 NOSIG= NOSIG=")

}
