% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmetar-package.R
\docType{data}
\name{metarWXcodes}
\alias{metarWXcodes}
\title{METAR WX weather conditions codes}
\format{
A data frame with 39 rows and 3 variables
}
\description{
A dataset containing the explanations for METAR WX
weather conditions codes. The variables are as follows:
}
\details{
\itemize{
\item Type; type of the codes
\item Abbreviation; the codes which are included in METAR reports
\item Meaning; description of the codes
}
}
\references{
\url{https://en.wikipedia.org/wiki/METAR#METAR_WX_codes}
}
\keyword{datasets}
