% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmetar-package.R
\docType{data}
\name{mst}
\alias{mst}
\title{Secondary airport list}
\format{
A character vector with the length of 9814 items
}
\description{
A character vector containing the list of airports.
}
\references{
\url{https://www.aviationweather.gov/docs/metar/stations.txt}
}
\author{
Greg Thompson from National Weather Service NCAR/RAP
}
\keyword{datasets}
