% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmetar-package.R
\docType{data}
\name{ourairports}
\alias{ourairports}
\title{Main list of airports}
\format{
A data frame with 28935 rows and 12 variables
}
\description{
#' A dataset containing the list of airports
}
\details{
From \url{https://ourairports.com}
A data set is in the public domain according to \url{https://ourairports.com/data/}

\itemize{
\item id; identification number
\item ident; airport ICAO code
\item type; airport type
\item name; airport name
\item latitude_deg; geographical latitude
\item longitude_deg; geographical longitude
\item elevation_ft; airport elevation in feet
\item elevation_m; airport elevation in meters
\item iso_country; ISO country code
\item iso_region; ISO region code
\item municipality;
\item iata_code; airport IATA code
}
}
\references{
\url{https://ourairports.com/data/}
}
\author{
David Megginson
}
\keyword{datasets}
