% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmetar-package.R
\docType{package}
\name{pmetar-package}
\alias{pmetar}
\alias{pmetar-package}
\title{pmetar: Processing METAR Weather Reports}
\description{
Processing METAR Weather Reports
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prcwiek/pmetar}
  \item Report bugs at \url{https://github.com/prcwiek/pmetar/issues}
}

}
\author{
\strong{Maintainer}: Pawel Cwiek \email{prc.altodato@gmail.com}

Other contributors:
\itemize{
  \item David Megginson (Author of data set with airports list https://ourairports.com/data/) [contributor]
  \item Greg Thompson (Author of data set with airports list https://weather.ral.ucar.edu/surface/stations.txt) [contributor]
}

}
\keyword{internal}
