% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_airport.R
\name{metar_airport}
\alias{metar_airport}
\title{Get airport ICAO, International Civil Aviation Organization, code.}
\usage{
metar_airport(x)
}
\arguments{
\item{x}{character vector; METAR weather report or reports.}
}
\value{
a character vector with an airport ICAO code.
}
\description{
Extract an airport ICAO code from METAR weather report.
}
\examples{
metar_airport("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_airport("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180")
metar_airport("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025")

}
