% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_day.R
\name{metar_day}
\alias{metar_day}
\title{Get day of month.}
\usage{
metar_day(x)
}
\arguments{
\item{x}{character vector; a METAR weather report or reports.}
}
\value{
a numeric vector with a day of a month.
}
\description{
Extract a day of a month from METAR weather report.
}
\examples{
metar_day("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_day("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180")
metar_day("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025")

}
