% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_decode.R
\name{metar_decode}
\alias{metar_decode}
\title{Decode METAR report.}
\usage{
metar_decode(x, metric = TRUE)
}
\arguments{
\item{x}{Input character vector}

\item{metric}{Selection between the metric system and the imperial system. As default metric = TRUE.}
}
\value{
A character vector or a tibble.
}
\description{
Function extracts information from METAR weather report. Depending of an input vector format
a returned data frame consists of different values.
}
\details{
For a current METAR report:\cr
- Airport ICAO\cr
- Day of Month\cr
- Hour (HH:mm)\cr
- Time zone\cr
- Wind speed \cr
- Wind speed unit (m/s or kn)\cr
- Gust \cr
- Gust unit (m/s or kn) \cr
- Wind shear\cr
- Wind direction (degrees)\cr
- Temperature (degrees Celsius)\cr
- Dew point (degrees Celsius)\cr
- Pressure (hPa)\cr
- Visibility\cr
- Visibility unit (m or miles)\cr
- Cloud coverage\cr
- Weather information from METAR WX codes\cr
- Airport Name\cr
- Longitude\cr
- Latitude\cr
- Elevation\cr
- Decode Date\cr
- Original METAR text\cr
- Source of information\cr
- Licence\cr\cr

For a historical METAR report:\cr
- Airport\cr
- MetarDate\cr
- Day of Month\cr
- Hour (HH:mm)\cr
- Time zone\cr
- Wind speed\cr
- Wind speed unit (m/s or kn)\cr
- Gust\cr
- Gust unit (m/s or kn)\cr
- Wind shear\cr
- Wind direction (degrees)\cr
- Temperature (degrees Celsius)\cr
- Dew point (degrees Celsius)\cr
- Pressure (hPa)\cr
- Visibility\cr
- Visibility unit (m or miles)\cr
- Cloud coverage\cr
- Weather information from METAR WX codes\cr
- Airport.name\cr
- Longitude\cr
- Latitude\cr
- Elevatio (m)\cr
- Decode.Date\cr
- Original METAR text\cr
- Source of information\cr
- Licence\cr
}
\examples{
metar_decode("EPWA 281830Z 18009KT 140V200 9999 SCT037 03/M01 Q1008 NOSIG")
metar_decode("CYUL 281800Z 13008KT 30SM BKN240 01/M06 A3005 RMK CI5 SLP180")
metar_decode("201711271930 METAR LEMD 271930Z 02002KT CAVOK 04/M03 Q1025 NOSIG= NOSIG=")

}
