% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_get.R
\name{metar_get}
\alias{metar_get}
\title{Get a current METAR reports or historical METAR reports.}
\usage{
metar_get(airport, start_date = "", end_date = "")
}
\arguments{
\item{airport}{Input character vector}

\item{start_date}{Input character vector}

\item{end_date}{Input character vector}
}
\value{
A character vector for a current METAR weather report.

A data frame character vectors with historical METAR weather report.
}
\description{
Function geta metar report from a web page.\cr
\cr
If only airport is passed as an argument, a current METAR weather report is donwloaded from\cr
the web page https://aviationweather.gov/metar/\cr
\cr
If a start date and an end date are specified, a set of historical METAR weather reports\cr
is is downloaded from the web page http://www.ogimet.com/
}
\examples{
metar_get("EPWA")
metar_get("CYUL")
metar_get("LEMD")
metar_get("EPWA", start_date = "2017-11-20", end_date = "2017-11-25")
metar_get("CYUL", start_date = "2016-07-01", end_date = "2016-07-10")
metar_get("LEMD", start_date = "2000-06-01", end_date = "2000-06-02")

}
