% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_get_historical.R
\name{metar_get_historical}
\alias{metar_get_historical}
\title{metar_get_historical}
\usage{
metar_get_historical(airport, start_date = "2017-11-21",
  end_date = "2017-11-22")
}
\arguments{
\item{end_date}{}
}
\description{
metar_get_historical
}
