% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_get_historical.R
\name{metar_get_historical}
\alias{metar_get_historical}
\title{Get historical METAR reports fro an airport.}
\usage{
metar_get_historical(
  airport = "EPWA",
  start_date = "2017-11-21",
  end_date = "2017-11-22",
  from = "iastate"
)
}
\arguments{
\item{airport}{Input character vector with an ICAO or IATA airport code}

\item{start_date}{Input character vector}

\item{end_date}{Input character vector}

\item{from}{Input character vector, allowed values are "iastate" and "ogimet"}
}
\value{
A character vector for a current METAR weather report.

A data frame character vectors with historical METAR weather report.
}
\description{
Function downloads a set of historical METAR weather reports.
The default source the Iowa Environmental Mesonet web page of Iowa State University
ASOS-AWOS-METAR http://mesonet.agron.iastate.edu/AWOS/ \cr
The secondary source is Weather Information Service provided by Ogimet http://www.ogimet.com/
}
\examples{
metar_get_historical("EPWA", start_date = "2017-11-20", end_date = "2017-11-25")
metar_get_historical("CYUL", start_date = "2016-07-01", end_date = "2016-07-10", from = "ogimet")
metar_get_historical("LEMD", start_date = "2000-06-01", end_date = "2000-06-02", from = "iastate")


}
