% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_location.R
\name{metar_location}
\alias{metar_location}
\title{Extract approximated airport location.}
\usage{
metar_location(x = "EPWA")
}
\arguments{
\item{x}{character vector; an airport ICAO four letters code or an IATA three letters code.}
}
\value{
A tibble with columns which consists of airport ICAO code, IATA code airport name, longitude, latitude,
 elevation in meters and source of information
}
\description{
Function findss approximated latitude, longitude and elevation of an airport according to \cr
IATA, International Air Transport Association, or \cr
ICAO, International Civil Aviation Organization, airport code
}
\examples{
metar_location("EPWA")
metar_location("CYUL")
metar_location("LEMD")
metar_location("NCRK")
metar_loaction("WAW")
metar_location("FRA")

}
