% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar_location.R
\name{metar_location}
\alias{metar_location}
\title{Extract approximated airport location.}
\usage{
metar_location(x)
}
\arguments{
\item{x}{Input character vector, an airport ICAO four letters code or IATA three letters code}
}
\value{
A tibble with columns which consists of airport ICAO code, IATA code airport name, longitude, latitude,
 elevation in meters, source of information
}
\description{
Function extract approximated lattitude, longitude and elevation for a given airport code.
}
\examples{
metar_location("EPWA")
metar_location("CYUL")
metar_location("LEMD")
metar_loaction("WAW")
metar_location("FRA")

}
