% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmetar-package.R
\docType{data}
\name{mst}
\alias{mst}
\title{Secondary airport list}
\format{
A character vector with the length of 10113 items
}
\description{
A character vector containing the list of airports.
}
\details{
From \url{https://www.aviationweather.gov/}
A data set is in the public domain according to \url{https://www.weather.gov/disclaimer/}
}
\references{
\url{https://weather.ral.ucar.edu/surface/stations.txt},
}
\author{
Greg Thompson from National Weather Service NCAR/RAP, NOAA National Weather Service
}
\keyword{datasets}
